+++ extensions/aws/data-plane-s3/src/test/java/org/eclipse/dataspaceconnector/aws/dataplane/s3/DataPlaneS3ExtensionTest.java
@@ -1,0 +1,41 @@
+/*
+ *  Copyright (c) 2022 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ */
+package org.eclipse.dataspaceconnector.aws.dataplane.s3;
+
+import org.eclipse.dataspaceconnector.dataplane.spi.pipeline.PipelineService;
+import org.eclipse.dataspaceconnector.junit.launcher.EdcExtension;
+import org.eclipse.dataspaceconnector.spi.types.domain.transfer.DataFlowRequest;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+
+import java.util.UUID;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+
+@ExtendWith(EdcExtension.class)
+class DataPlaneS3ExtensionTest {
+
+    @Test
+    void shouldProvidePipelineServices(PipelineService pipelineService) {
+        var request = DataFlowRequest.Builder.newInstance()
+                .processId(UUID.randomUUID().toString())
+                .sourceDataAddress(TestFunctions.validS3DataAddress())
+                .destinationDataAddress(TestFunctions.validS3DataAddress())
+                .build();
+
+        var result = pipelineService.validate(request);
+
+        assertThat(result.succeeded()).isTrue();
+    }
+}