+++ extensions/aws/s3/s3-core/src/main/java/org/eclipse/dataspaceconnector/aws/s3/core/S3CoreExtension.java
@@ -1,0 +1,31 @@
+/*
+ *  Copyright (c) 2022 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ */
+package org.eclipse.dataspaceconnector.aws.s3.core;
+
+import org.eclipse.dataspaceconnector.spi.system.Provides;
+import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
+import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
+
+@Provides({ S3ClientProvider.class })
+public class S3CoreExtension implements ServiceExtension {
+    @Override
+    public String name() {
+        return "S3";
+    }
+
+    @Override
+    public void initialize(ServiceExtensionContext context) {
+        context.registerService(S3ClientProvider.class, new S3ClientProviderImpl());
+    }
+}