+++ CHANGELOG.md
@@ -31,6 +31,7 @@ the detailed section referring to by linking pull requests or issues.
 * Add configuration setting for state machine batch size (#872)
 * Add Jetty context alias for IDS API (#815)
 * Add `Hostname` service (#948)
+* Implement S3 data source and data sink for data-plane (#887)
 * Add embedded and remote DPF Selector (#832)
 * Pass path information into http data source through DPF public API (#929)
 * SQL-based TransferProcessStore (#865)